full-stack updated
I restructured everything, organized most of the project, small change in database exist also, and now the routes are available and ready to work on. however I couldn't go deep dive to the details but it's well structured.
I recommend the first thing to start working on to be the register, that lawyers won't get session when they gor created, due some errors mostly because the education part, which I believe if you check the view you get the idea how the backend must work.

and other routes must be double checked and interested with the front-end